refactor: types declaration
