+++ package.json
@@ -5,7 +5,7 @@
   "main": ".build/index.js",
   "type": "module",
   "scripts": {
-    "build": "tsc",
+    "build": "tsc --project tsconfig.build.json",
     "start": "pm2 start dist/index.js --name next-blog-api --node-args=\"--no-warnings --experimental-specifier-resolution=node\"",
     "monit": "pm2 monit",
     "stop": "pm2 stop next-blog-api",
